Fixing changes watch
It's a little too complex now, but that's the best idea so far.

We have a 'father' thread which has a child watching the changes. The
child is set as deamon, so once the father dies (because of an event)
the child dies too, and kills the connection to the continuous changes
feed.