Fix #641
This commit fixes two problems:
* ticket #641
* (no ticket): The change watch was stopping after a couple of seconds
since the last change.

The first problem was fixed using the correct seq number once all the
design documents, so the workspace document is not a change anymore.
Also, a try/catch was added to loadChange inside the ChangeController so
we don't kill the change watcher thread when an exception is raised

The second problem was fixed adding a heartbeat to the change stream.
It's weird, but it seems that after a couple of seconds of inactivity
the stream doesn't work anymore.