CSV export directive refactor
Previous CSVExporter is now called fileExporter.
The directive is now generic, allowing all files to be exported.
Content must be generated from controller as well as file name, extension and type.
All this fields must be provided to the directive.
No sanitization is performed on the content. So be careful.
Test #810