Refactor filterInaccesibleWorkspaces() function
The function now filters appropriately. For every workspace name
retrieved from the DB it checks whether the user has privileges and if
not, it filters it from the list.
Also, since the workspaces service now returns a promise instead of a
list of workspaces, the statusReport service had to be modified to
reflect this change.

Test #895