Add modal controllers and partials to Hosts comp
Modified statusReportCtrl to use new hostsManager provider. The new
provider has a specific function to maintain compatibility with the
Status Report component. In the future the Status Report should be able
to support async requests.
Added modalDelete and modalKO to commons component. These should be the
ones used in all other components and not the ones on partials/ or
modal/.
Added 'hosts' route. No links to it yet.