Refactor components, mainly Status Report
Status Report used to have some of its partials and controllers (mainly
for modals) in a common directory which didn't make sense, since they
only made sense to this component. In this commit I started to
separate the files either putting them inside Status Report or
moving them to the commons component.

Accepted #1121