Refactor D3 Directive for horizontal stacked bar
Changed directive name to reflect its behaviour, changed the name in
partial as well.
Added functionality to VulnsByPrice controller, it now calculates the prices
based on the vulnerabilities available for the current workspace.
Added two methods to dashboard service; one to get all vulnerabilities and one to
accumulate prices in an array of objects. Stole this method from
summarized controller. TODO: modify summarized controller to use the one
on the service.