Fix getVulns method in vulnsManager
The promise that this method was returning was the one received from
$http.get(ws), which means that when that request was fulfilled the
promise was resolved, but vulnsManager.vulns wasn't ready yet. The only
post-condition for this method is that when the returned promised is
resolved vulnsManager.vulns must be ready.
Also, modified vulnsManager._load() to return the array, because why
not.