Fix several bugs in Vuln and WebVuln
First of all, both constructors now receive an extra param containing WS
name, this is to avoid having to set the WS inside vuln data when passed
around.
Also, both objects now set properties only if not undefined in data, so
fields will persist when a partial update is performed.