Fix license edit modal bug when leaving empty fields
Also, fixed the validation. There seems to be a problem when setting
datepicker input fields as required; the value of the fields disappears
and they are set as undefined. Couldn't solve that problem so I removed
the required property and added a manual validation which sucks. This
should be solved asap.