Fix bug when creating a template manually
When the Vuln Template was created manually, the "description" field was
lost. Fixed it so it gets saved in the "description" field instead of
the desc_summary.
Also modified the list view so that it prints the desc_summary field in
case the description is empty.
Also added desc_summary as a property of the object, which for some
reason wasn't there already.