Fix vuln template duplication
The vulnerability templates are queried by the user from the
vulnerability creation dialog in the web UI. The only field the user has
to pick one template over another is the name. Therefore, having
more than one vuln template with the same name doesn't make sense. The
new model doesn't allow duplication, but the old one did. The importer
is now able to detect duplicated template names and appends "(n)" to the
name with n being the number of repetition.