Add initial working code for commands stats
This commit solves the historic stats for commands.
When a command creates a vuln, host, setvice etc a flag will be added on
the results.
When another command uses an already created object the flag will be at
0.
This commit has the following:
    * Add columns fields to generate created information. when a command
creates the object a field with 1 will be returned in the columns with
name: created_*.
    * On the command model sum_created_* will have the value of the sum of
all created object.
    * Added a test to verify a simple case.