Add generic write tests for credentials API (tests success story)
Testing success story: I found a bug when updating a host credential and
changing its parent to a service. It was trying to save a credential
with both a host and a service, and the check constraint failed. This
happen because host_id wasn't set to None.

The downside of this is that the bug only happened some times because
the credential factory randomly choices if its parent is a host or a
service.