Improve validate_uniqueness overriding in vulns API
It was setting attributes in the class instance that depended on a
property of the request. This is a bad pattern, since view instances are
reused and not instanced once per request.

Tests success story: I first forgot to replace `self.unique_fields` for
`unique_fields` in the for loop. Before committing I ran the tests and
found that the conflicts on creation were not happening, and I noted
that I almost introduce a regression in the code, but the test cases
catched me before doing it.