Fix vuln web with host parent error message of the web ui.
In integration, when the user set type to vuln web and selected one or
more hosts as target, a message "Error: Trying to create a Web
Vulnerability with one or more Host parents. Please deselect the Host
targets or click here to deselect all targets" was shown. The migration
changes broke it because the api didn't return the object type for host
and services. It is fixed now.

Another victory of undefined, that made the browser don't throw any
exceptions or warnings.