Refactor to fix race condition
We detected that when the server was running with multiple threads or
processes a race condition could be triggered when we first select
object from the database.
With these changes we try to insert data first and if the unique
constraint failed then we use unique fields + data to select the
conflicting object from the database.
With this we return the object which will be later used for conflict
resolution