Include server/www in the build
So the path is not hardcoded anymore. I had to use a MANIFEST.in
instead of package_data to recursively include directories. I had
to spend a lot of time debugging why it wasn't working, and it turned
out to be that I had to prepend a ./ in the path. This is an ugly
setuptools bug.