Remove flask shell hack not needed anymore
It was an ugly hack to make `flask shell` work before we had a package.
Now it isn't needed, and it can make some tests that should break pass
silently, so it is necessary to remove this piece of code.

To get a shell now, you can run `FLASK_APP=faraday.server.web flask
shell`