Update readme. Fix several links. Remove old quickstart
Several users tried to install faraday from sourcecode and this
causes issues on installation since they were doing a dev install.

Changed the readme to let the users known that the proper way to install
faraday for final usage is to use the packages that we provide.

Fixed several broken and old links in the wiki