Fix version of faraday-server Nix derivation
Previously, it was using the version indicated by pynixify. If we
changed faraday/__init__.py and didn't run pynixify again, the version
indicated in the derivation name would be stale.

This change overrides the version by directly reading
faraday/__init__.py in the Nix code. This way, there is no need to
re-run pynixify to refresh the version. This will be done automatically.